feat: move session listing to experimental routes with enhanced query parameters
